adds reprocess option to Add On Demand Menu in scene-detail component on Custom PRoducts
