+++ src/app/components/results-menu/scenes-list/scene/scene.component.ts
@@ -1,8 +1,6 @@
 import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
 import { faCopy } from '@fortawesome/free-solid-svg-icons';
 
-import { ConfirmationComponent } from '@components/header/processing-queue/confirmation/confirmation.component';
-
 import * as services from '@services';
 import * as models from '@models';
 
@@ -10,11 +8,7 @@ import { QueuedHyp3Job, Hyp3ableProductByJobType } from '@models';
 import { AppState } from '@store';
 import { Store } from '@ngrx/store';
 
-import * as hyp3Store from '@store/hyp3';
 import * as queueStore from '@store/queue';
-import { MatDialog } from '@angular/material/dialog';
-import { catchError, finalize } from 'rxjs/operators';
-import { of } from 'rxjs';
 
 @Component({
   selector: 'app-scene',
@@ -44,31 +38,16 @@ export class SceneComponent implements OnInit {
   public copyIcon = faCopy;
   public SearchTypes = models.SearchType;
 
-  public projectName = '';
-  public processingOptions: models.Hyp3ProcessingOptions;
-  public validateOnly = false;
-
   constructor(
     public env: services.EnvironmentService,
     private screenSize: services.ScreenSizeService,
     private hyp3: services.Hyp3Service,
     private store$: Store<AppState>,
-    public dialog: MatDialog,
-    private notificationService: services.NotificationService,
   ) { }
 
   ngOnInit(): void {
     this.screenSize.breakpoint$.subscribe(
       breakpoint => this.breakpoint = breakpoint
-    );
-
-
-    this.store$.select(hyp3Store.getProcessingOptions).subscribe(
-      options => this.processingOptions = options
-    );
-
-    this.store$.select(hyp3Store.getProcessingProjectName).subscribe(
-      projectName => this.projectName = projectName
     );
   }
 
@@ -79,7 +58,6 @@ export class SceneComponent implements OnInit {
   public onClearFocused(): void {
     this.hovered = false;
   }
-
 
   public withOffset(val: number, offset: number): number {
     return Math.trunc(val + offset);
@@ -134,13 +112,13 @@ export class SceneComponent implements OnInit {
     let byProductType: models.Hyp3ableByProductType[] = [];
     byProductType.push(temp);
 
-    let hp = {
+    const hyp3ableProduct = {
       byProductType,
       total: 1,
       jobType: job_types[job_type]
     } as models.Hyp3ableProductByJobType
-    byJobType.push(hp);
-    // byJobType.push(byProductType);
+
+    byJobType.push(hyp3ableProduct);
 
     const output = {
       byJobType,
@@ -148,109 +126,6 @@ export class SceneComponent implements OnInit {
     } as {byJobType: models.Hyp3ableProductByJobType[], total: number}
 
     return output;
-  }
-
-  public onSubmitQueue(jobTypesWithQueued, validateOnly: boolean) {
-    console.log(jobTypesWithQueued);
-    console.log(validateOnly);
-
-    console.log(this.processingOptions);
-
-    const processOptionKeys = Object.keys(this.scene.metadata.job.job_parameters).filter(key => key !== 'granules');
-    let processingOptions = {};
-    processOptionKeys.forEach(key => processingOptions[key] = this.scene.metadata.job.job_parameters[key]);
-    // let processingOptions = {[key in processOptionKeys]: this.scene.metadata.job[key]};
-    const hyp3JobsBatch = this.hyp3.formatJobs(jobTypesWithQueued, {
-      projectName: this.projectName,
-      processingOptions
-    });
-
-    this.hyp3.submiteJobBatch$({jobs: hyp3JobsBatch, validate_only: false}).pipe(
-      catchError(resp => {
-        if (resp.error) {
-          if (resp.error.detail === 'No authorization token provided' || resp.error.detail === 'Provided apikey is not valid') {
-            this.notificationService.error('Your authorization has expired. Please sign in again.', 'Error', {
-              timeOut: 5000,
-          });
-          } else {
-            this.notificationService.error( resp.error.detail, 'Error', {
-              timeOut: 5000,
-            });
-          }
-        }
-
-        return of({jobs: null});
-      }),
-      finalize(() => {
-        this.store$.dispatch(new hyp3Store.LoadUser());
-      }),
-    ).subscribe(
-      (resp: any) => {
-        if (resp.jobs === null) {
-          return;
-        }
-        console.log(resp);
-
-        // const successfulJobs = resp.jobs.map(job => ({
-        //   granules: job.job_parameters.granules.map(g => ({name: g})),
-        //   job_type: models.hyp3JobTypes[job.job_type]
-        // }));
-
-        // this.store$.dispatch(new queueStore.RemoveJobs(successfulJobs));
-      }
-    );
-    // const hyp3JobsBatch = this.hyp3.formatJobs(jobTypesWithQueued, {
-    //   projectName: this.projectName,
-    //   processingOptions: this.processingOptions
-  // });
-  }
-
-  public onReviewQueue() {
-
-    const job_types = models.hyp3JobTypes;
-    const job_type = Object.keys(job_types).find(id =>
-      {
-        return this.scene.metadata.job.job_type === id as any;
-      });
-
-      // this.scene.metadata.job.job_parameters
-
-    const confirmationRef = this.dialog.open(ConfirmationComponent, {
-      id: 'ConfirmProcess',
-      width: '350px',
-      height: '500px',
-      maxWidth: '350px',
-      maxHeight: '500px',
-      data: [{
-        jobType: job_types[job_type],
-        selected: true,
-        jobs: [{
-          granules: this.scene.metadata.job.job_parameters.scenes,
-          job_type: job_types[job_type]
-        } as QueuedHyp3Job ]
-      }]
-    });
-
-    confirmationRef.afterClosed().subscribe(
-      jobTypesWithQueued => {
-        if (!jobTypesWithQueued) {
-          return;
-        }
-
-        if (this.env.maturity === 'prod') {
-          this.validateOnly = false;
-        }
-
-        this.onSubmitQueue(
-          jobTypesWithQueued,
-          this.validateOnly
-        );
-      }
-    );
-  }
-
-  public onValidateOnlyToggle(val: boolean): void {
-    this.validateOnly = val;
   }
 
   public isExpired(job: models.Hyp3Job): boolean {
