+++ src/app/components/shared/on-demand-add-menu/on-demand-add-menu.component.html
@@ -39,7 +39,7 @@
         [disabled]="!expiredJobs"
         (click)="onReviewQueue()"
         mat-menu-item>
-        Re-Process Job
+        Re-Process Job ({{remaining}} remaining this month)
 
         <mat-icon
           class="clickable" style="margin-left: 10px;">
